fix: address review feedback for paste image context menu
Inline clipboardUtil into useImageMenuOptions (sole consumer), derive
file extension from MIME type, and call both pasteFile and pasteFiles
for consistency with usePaste.ts.
